requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.9.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
